+++ src/Util/Lifecycle/Method/PaymentMethodDataRegistry.php
@@ -37,7 +37,6 @@ class PaymentMethodDataRegistry
         MyBankMethodData::class,
         OxxoMethodData::class,
         P24MethodData::class,
-        SofortMethodData::class,
         TrustlyMethodData::class,
         VenmoMethodData::class,
         PayLaterMethodData::class,