+++ tests/OrdersApi/Builder/APMOrderBuilderTest.php
@@ -24,7 +24,6 @@ use Swag\PayPal\OrdersApi\Builder\APM\MultibancoOrderBuilder;
 use Swag\PayPal\OrdersApi\Builder\APM\MyBankOrderBuilder;
 use Swag\PayPal\OrdersApi\Builder\APM\OxxoOrderBuilder;
 use Swag\PayPal\OrdersApi\Builder\APM\P24OrderBuilder;
-use Swag\PayPal\OrdersApi\Builder\APM\SofortOrderBuilder;
 use Swag\PayPal\OrdersApi\Builder\APM\TrustlyOrderBuilder;
 use Swag\PayPal\OrdersApi\Builder\Util\AddressProvider;
 use Swag\PayPal\OrdersApi\Builder\Util\AmountProvider;
@@ -41,7 +40,6 @@ use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Multibanco;
 use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\MyBank;
 use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Oxxo;
 use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\P24;
-use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Sofort;
 use Swag\PayPal\RestApi\V2\Api\Order\PaymentSource\Trustly;
 use Swag\PayPal\Setting\Settings;
 use Swag\PayPal\Test\Helper\ConstantsForTesting;
@@ -214,7 +212,6 @@ class APMOrderBuilderTest extends TestCase
             [MyBankOrderBuilder::class, [], MyBank::class, []],
             [OxxoOrderBuilder::class, [], Oxxo::class, ['email' => 'test@test.com']],
             [P24OrderBuilder::class, [], P24::class, ['email' => 'test@test.com']],
-            [SofortOrderBuilder::class, [], Sofort::class, []],
             [TrustlyOrderBuilder::class, [], Trustly::class, []],
         ];
     }