+++ tests/Util/Lifecycle/State/PaymentMethodStateServiceTest.php
@@ -41,7 +41,6 @@ class PaymentMethodStateServiceTest extends TestCase
         'Swag\PayPal\Checkout\Payment\Method\MyBankAPMHandler' => false,
         'Swag\PayPal\Checkout\Payment\Method\OxxoAPMHandler' => true,
         'Swag\PayPal\Checkout\Payment\Method\P24APMHandler' => true,
-        'Swag\PayPal\Checkout\Payment\Method\SofortAPMHandler' => false,
         'Swag\PayPal\Checkout\Payment\Method\TrustlyAPMHandler' => true,
         'Swag\PayPal\Checkout\Payment\Method\VenmoHandler' => false,
         'Swag\PayPal\Checkout\Payment\Method\PayLaterHandler' => true,